(crazy squash) replaced malloc/free with macros
because nvcc is really C++, we need casts around
every void* in the code.
instead of adding those explicitly, just made
the malloc/free stuff macros that will take
care of a bit more type safety.
in fact, this caught a bad function declaration
(ulongs_subset).

+++ subset.h
@@ -7,7 +7,7 @@ unsigned* uints_subset(
     unsigned const* a,
     unsigned const* offsets);
 
-unsigned* ulongs_subset(
+unsigned long* ulongs_subset(
     unsigned n,
     unsigned width,
     unsigned long const* a,
